requirements.txt
aniso8601==9.0.1
beautifulsoup4==4.11.1
datefinder==0.7.3
fuzzywuzzy==0.18.0
JPype1==1.4.1
kafka==1.3.5
kafka_python==2.0.2
nltk==3.6.1
numpy==1.20.1
pandas==1.2.4
pytest==6.2.3
python_dateutil==2.8.2
requests==2.28.1
spacy==3.4.1
dateparser==1.1.1
scikit_learn==1.2.2
black==19.10.0
pylint==2.7.4

//////////////////////////////////////////
annotated-types==0.6.0
apturl==0.5.2
bcrypt==3.2.0
beautifulsoup4==4.12.2
blinker==1.4
blis==0.7.11
Brlapi==0.8.3
bs4==0.0.1
catalogue==2.0.10
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
cloudpathlib==0.16.0
colorama==0.4.4
command-not-found==0.3
confection==0.1.4
cryptography==3.4.8
cupshelpers==1.0
cymem==2.0.8
datefinder==0.7.3
dateparser==1.2.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.1
duplicity==0.8.21
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1-py3-none-any.whl
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
fasteners==0.14.1
future==0.18.2
fuzzywuzzy==0.18.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
Jinja2==3.1.3
joblib==1.3.2
JPype1==1.5.0
kafka==1.3.5
kafka-python==2.0.2
keyring==23.5.0
langcodes==3.3.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
murmurhash==1.0.10
netifaces==0.11.0
nltk==3.8.1
numpy==1.26.3
oauthlib==3.2.0
olefile==0.46
packaging==23.2
pandas==2.1.4
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
preshed==3.0.9
protobuf==3.12.4
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
pydantic==2.5.3
pydantic_core==2.14.6
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
regex==2023.12.25
reportlab==3.6.8
requests==2.25.1
scikit-learn==1.4.0
scipy==1.11.4
SecretStorage==3.3.1
simplejson==3.17.6
six==1.16.0
smart-open==6.4.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sutime==1.0.1
systemd-python==234
textblob==0.17.1
thinc==8.2.2
threadpoolctl==3.2.0
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
wasabi==1.1.2
weasel==0.3.4
xdg==5
xkit==0.0.0
zipp==1.0.0
